+++ miden-vm/tests/integration/operations/decorators/advice.rs
@@ -110,7 +110,10 @@ fn advice_push_mapval() {
 
     let test = build_test!(source, &stack_inputs, [], MerkleStore::default(), adv_map);
     test.expect_stack(&[5]);
+}
 
+#[test]
+fn adv_push_mapvaln() {
     // --- test simple adv.push_mapvaln --------------------------------------------
     let source: &str = "
     begin
@@ -134,6 +137,93 @@ fn advice_push_mapval() {
 
     let test = build_test!(source, &stack_inputs, [], MerkleStore::default(), adv_map);
     test.expect_stack(&[15, 14, 13, 12, 11, 5]);
+}
+
+#[test]
+fn adv_push_mapvaln_padding() {
+    // --- test adv.push_mapvaln.0 -------------------------------------------------
+    let source: &str = "
+    begin
+        # stack: [4, 3, 2, 1, ...]
+
+        # load the advice stack with values from the advice map (including the number
+        # of elements) and drop the key
+        # since 0 was provided as an immediate value, the resulting map values should not be padded
+        adv.push_mapvaln.0
+        dropw
+
+        # move the values from the advice stack to the operand stack
+        adv_push.6
+        swapdw dropw dropw
+    end";
+
+    let stack_inputs = [1, 2, 3, 4];
+    let adv_map = [(
+        Word::try_from(stack_inputs).unwrap(),
+        vec![Felt::new(11), Felt::new(12), Felt::new(13), Felt::new(14), Felt::new(15)],
+    )];
+
+    let test = build_test!(source, &stack_inputs, [], MerkleStore::default(), adv_map);
+    test.expect_stack(&[15, 14, 13, 12, 11, 5]);
+
+    // --- test adv.push_mapvaln.4 -------------------------------------------------
+    let source: &str = "
+    begin
+        # stack: [4, 3, 2, 1, ...]
+
+        # load the advice stack with values from the advice map (including the number
+        # of elements) and drop the key
+        # since 4 was provided as an immediate value, the resulting map values should be padded to
+        # the next multiple of 4
+        adv.push_mapvaln.4
+        dropw
+
+        # move the values from the advice stack to the operand stack
+        adv_push.5
+        swapdw dropw dropw
+    end";
+
+    let stack_inputs = [1, 2, 3, 4];
+    let adv_map = [(
+        Word::try_from(stack_inputs).unwrap(),
+        vec![Felt::new(11), Felt::new(12), Felt::new(13)],
+    )];
+
+    let test = build_test!(source, &stack_inputs, [], MerkleStore::default(), adv_map);
+    test.expect_stack(&[0, 13, 12, 11, 3]);
+
+    // --- test adv.push_mapvaln.8 -------------------------------------------------
+    let source: &str = "
+    begin
+        # stack: [4, 3, 2, 1, ...]
+
+        # load the advice stack with values from the advice map (including the number
+        # of elements) and drop the key
+        # since 8 was provided as an immediate value, the resulting map values should be padded to
+        # the next multiple of 8
+        adv.push_mapvaln.8
+        dropw
+
+        # move the values from the advice stack to the operand stack
+        adv_push.8 swapdw dropw dropw
+        adv_push.1 movup.9 drop
+    end";
+
+    let stack_inputs = [1, 2, 3, 4];
+    let adv_map = [(
+        Word::try_from(stack_inputs).unwrap(),
+        vec![
+            Felt::new(11),
+            Felt::new(12),
+            Felt::new(13),
+            Felt::new(14),
+            Felt::new(15),
+            Felt::new(16),
+        ],
+    )];
+
+    let test = build_test!(source, &stack_inputs, [], MerkleStore::default(), adv_map);
+    test.expect_stack(&[0, 0, 16, 15, 14, 13, 12, 11, 6]);
 }
 
 #[test]